fix: lots of fixes only "tovalidate" not up

+++ EJBank-ejb/src/main/java/com/ejbank/service/user/UserServiceLocal.java
@@ -14,7 +14,7 @@ public interface UserServiceLocal {
 
     AccountsDto getAccounts(Integer id);
 
-    Integer getNotificationCount(Integer id);
+    Long getNotificationCount(Integer id);
 
     AccountsWithUserDto getAccountsWithUser(Integer id);
